Fixed problems with Merged code and added the current approval status to page
