--- src/SFA.DAS.EAS.Web/Views/EmployerCommitments/TransferApproval.cshtml
@@ -1,127 +1,0 @@
-﻿@using SFA.DAS.EmployerCommitments.Web
-@using SFA.DAS.EmployerCommitments.Web.Extensions
-@using SFA.DAS.EmployerCommitments.Web.ViewModels
-
-@{
-    ViewBag.Title = "Transfer request details";
-    ViewBag.Section = "apprentices";
-    ViewBag.PageID = "approve-transfer-cohort";
-    ViewBag.Question = "Do you want to approve this transfer request?";
-}
-
-@model OrchestratorResponse<TransferCommitmentViewModel>
-
-
-<div class="grid-row">
-    <div class="column-two-thirds">
-        <form method="POST" action="@Url.Action("TransferApproval")" onsubmit="sfa.tagHelper.submitRadioForm('@ViewBag.Title')">
-            
-            @Html.Partial("ValidationSummary", Html.ViewData.ModelState)
-            @Html.AntiForgeryToken()
- 
-            <div class="form-group">
-                <legend>
-                    <h1 class="heading-xlarge">@ViewBag.Title</h1>
-                </legend>
-            </div>
-
-            <div class="grid-row">
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Account Name
-                    </div>
-                    <div class="column-two-thirds">
-                        @Model.Data.LegalEntityName
-                    </div>
-                </div>
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Account Id
-                    </div>
-                    <div class="column-two-thirds">
-                        @Model.Data.HashedTransferReceiverAccountId
-                    </div>
-                </div>
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Cohort Reference
-                    </div>
-                    <div class="column-half">
-                        @Model.Data.HashedCohortReference
-                    </div>
-                </div>
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Status
-                    </div>
-                    <div class="column-two-thirds">
-                        Pending
-                    </div>
-                </div>
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Total Cost
-                    </div>
-                    <div class="column-two-thirds">
-                        @($"£{Model.Data.TotalCost:n0}")
-                    </div>
-                </div>
-                <div class="column-full">
-                    <hr class="hr-top" />
-                    <div class="column-one-third">
-                        Training
-                    </div>
-                    <div class="column-two-thirds">
-                        @foreach (var item in Model.Data.TrainingList)
-                        {
-                            <p>@item.SummaryDescription</p>
-                        }
-                    </div>
-                    <hr class="hr-bottom" />
-                </div>
-            </div>
-            
-
-            <fieldset role="radiogroup" aria-required="true">
-                <div class="form-error-group form-group @Html.AddClassIfPropertyInError("ApprovalConfirmed", "error")">
-                    <h1 class="heading-medium" for="">@ViewBag.Question</h1>
-                    <div>
-                        @Html.ValidationMessage("ApprovalConfirmed", "Please choose an option", new { id = "error-message-Confirmation" })
-                    </div>
-                    <label class="block-label selection-button-radio" for="ApprovalConfirmed-True">
-                        @Html.RadioButton("ApprovalConfirmed", "True", new { id = "Confirmation-True", dataOptionName = "Yes", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Question}', 'Yes')" })
-                        Yes, use approve the request
-                    </label>
-                    <label class="block-label selection-button-radio" for="ApprovalConfirmed-False">
-                        @Html.RadioButton("ApprovalConfirmed", "False", new { id = "Confirmation-False", dataOptionName = "No", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Question}', 'No')" })
-                        No, reject the request
-                    </label>
-                </div>
-            </fieldset>
-            
-            <button type="submit" class="button" id="submit-transfer-connection">Continue</button>
-            <a class="links" href="@Url.Action("Index", "EmployerCommitments")" aria-label="Cancel">Cancel</a>
-            
-            @Html.Hidden("HashedTransferSenderAccountId", Model.Data.HashedTransferSenderAccountId)
-            @Html.Hidden("HashedTransferReceiverAccountId", Model.Data.HashedTransferReceiverAccountId)
-            @Html.Hidden("HashedCommitmentId", Model.Data.HashedCohortReference)
-
-        </form>
-    </div>
-</div>
-@section breadcrumb {
-    <div class="breadcrumbs">
-        <a href="@Url.Action("Index", "EmployerCommitments")" aria-label="Back to Apprentices" class="back-link">Back to Apprentices</a>
-    </div>
-}
-
-@section gaDataLayer {
-    <script>
-        sfa.dataLayer.vpv = '/accounts/apprentices/legalEntity/create/choose-organisation';
-    </script>
-}
