+++ src/SFA.DAS.EAS.Web/Views/Transfers/TransferApproval.cshtml
@@ -13,7 +13,7 @@
 
         <h1 class="heading-xlarge">@ViewBag.Title</h1>
    
-    @Html.AntiForgeryToken()
+   
 
 
 <div class="grid-row">
@@ -32,7 +32,7 @@
             Account Id
         </div>
         <div class="column-two-thirds">
-            @Model.Data.HashedAccountId
+            @Model.Data.HashedTransferReceiverAccountId
         </div>
     </div>
     <div class="column-full">
@@ -50,7 +50,7 @@
             Status
         </div>
         <div class="column-two-thirds">
-            Pending
+            @Model.Data.TransferApprovalStatus
         </div>
     </div>
     <div class="column-full">
@@ -79,30 +79,30 @@
 
 <div class="grid-row">
     <div class="column-two-thirds">
-        <form method="POST" action="@Url.Action("ApproveTransfer")" onsubmit="sfa.tagHelper.submitRadioForm('@ViewBag.Title')">
-            
+        <form method="POST" action="@Url.Action("TransferApproval")" onsubmit="sfa.tagHelper.submitRadioForm('@ViewBag.Title')">
+
             @Html.Partial("ValidationSummary", Html.ViewData.ModelState)
 
             <fieldset role="radiogroup" aria-required="true">
                 <div class="form-error-group form-group @Html.AddClassIfPropertyInError("Confirmation", "error")">
                     <h2 class="heading-medium" for="">Do you want to approve this transfer request?</h2>
                     <div>
-                        @Html.ValidationMessage("Confirmation", "Please choose an option", new { id = "error-message-Confirmation" })
+                        @Html.ValidationMessage("ApprovalConfirmed", "Please choose an option", new { id = "error-message-Confirmation" })
                     </div>
                     <label class="block-label selection-button-radio" for="Confirmation-True">
-                        @Html.RadioButton("Confirmation", "True", new { id = "Confirmation-True", dataOptionName = "Yes", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Title}', 'Yes')" })
+                        @Html.RadioButton("ApprovalConfirmed", "True", new { id = "Confirmation-True", dataOptionName = "Yes", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Title}', 'Yes')" })
                         Yes, use approve the request
                     </label>
                     <label class="block-label selection-button-radio" for="Confirmation-False">
-                        @Html.RadioButton("Confirmation", "False", new { id = "Confirmation-False", dataOptionName = "No", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Title}', 'No')" })
+                        @Html.RadioButton("ApprovalConfirmed", "False", new { id = "Confirmation-False", dataOptionName = "No", onclick = $"sfa.tagHelper.radioButtonClick('{ViewBag.Title}', 'No')" })
                         No, reject the request
                     </label>
                 </div>
             </fieldset>
-            
             <button type="submit" class="button" id="submit-transfer-connection">Continue</button>
             <a class="button-link" href="@Url.Action("Index", "EmployerCommitments")" aria-label="Cancel">Cancel</a>
-
+            @Html.AntiForgeryToken()
+            @Html.Hidden("HashedTransferReceiverAccountId", Model.Data.HashedTransferReceiverAccountId)
         </form>
     </div>
 </div>
